+++ tests/nws/products/test_nldn.py
@@ -2,17 +2,21 @@
 
 import pytest
 from pyiem.nws.products.nldn import parser
-from pyiem.util import get_test_file
+from pyiem.util import get_test_filepath
 
 
 def test_1_basic():
     """CLIBNA is a new diction"""
-    np = parser(get_test_file("NLDN/example.bin", fponly=True))
+    fp = get_test_filepath("NLDN/example.bin")
+    with open(fp, "rb") as fh:
+        np = parser(fh)
     assert len(np.df.index) == 50
 
 
 @pytest.mark.parametrize("database", ["nldn"])
 def test_sql(dbcursor):
     """Test that we can insert data."""
-    np = parser(get_test_file("NLDN/example.bin", fponly=True))
+    fp = get_test_filepath("NLDN/example.bin")
+    with open(fp, "rb") as fh:
+        np = parser(fh)
     np.sql(dbcursor)